seperate LDAPFunktions
